made changes to mock layout
