--- client/Questions/AnswersList.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const AnswersList = (props) => {
-  return (
-    <div>This is an answer</div>
-  )
-}
-
-export default AnswersList;
