--- client/Questions/QuestionEntry.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const QuestionEntry = (props) => {
-  return (
-    <div>This is a question</div>
-  )
-}
-
-export default QuestionEntry;
