+++ client/components/App.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import RatingsApp from './ratings/RatingsApp.js'
+import QuestionsList from './Questions/QuestionsList.jsx';
 
 const App = (props) => (
   <div>
@@ -9,6 +10,9 @@ const App = (props) => (
     </div>
     <div id="reviews-ratings">
       <RatingsApp />
+    </div>
+    <div id="questions_answers">
+      <QuestionsList />
     </div>
   </div>
 )
