+++ client/components/Questions/QuestionEntry.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const QuestionEntry = (props) => {
+  return (
+    <div>
+      <span className='question'> Q: This is a question</span>
+      <span className='question-helpful'>Helpful? Yes(count) | Add Answer</span>
+    </div>
+  )
+}
+
+export default QuestionEntry;
